add and apply clubDetail page component
