+++ components/pages/index.ts
@@ -2,10 +2,12 @@ import Home from './Home';
 import Profile from './Profile';
 import CreateClub from './CreateClub';
 import Login from './Login';
+import ClubDetail from './ClubDetail';
 
 export default {
   Home,
   Profile,
   CreateClub,
-  Login
+  Login,
+  ClubDetail
 };
