+++ pages/clubs/[id].tsx
@@ -1,3 +1,4 @@
+import Pages from '@/components/pages';
 import Templates from '@/components/templates';
 import useCategories from '@/hooks/useCategories';
 import useClubDetail from '@/hooks/useClubDetail';
@@ -8,16 +9,10 @@ import { QueryClient, dehydrate, useQuery } from '@tanstack/react-query';
 import { GetServerSidePropsContext, NextPage } from 'next';
 import { useRouter } from 'next/router';
 
-export async function getServerSideProps(context: GetServerSidePropsContext) {
-  const { params } = context;
-  const clubId = params?.id ? Number(params.id) : -1;
+export async function getServerSideProps() {
   const queryClient = new QueryClient();
 
   await queryClient.prefetchQuery(['auth'], AuthService.googleLogin);
-  await queryClient.prefetchQuery(
-    ['clubDetail', clubId],
-    ClubService.getDetail
-  );
 
   return {
     props: {
@@ -27,28 +22,13 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
 }
 
 const Club: NextPage = () => {
-  const { query } = useRouter();
-  const clubId = query?.id ? Number(query?.id) : -1;
+  const { isLoading, isError } = useNavigationGuard();
 
-  const guard = useNavigationGuard();
-  const categories = useCategories();
-  const clubDetail = useClubDetail(clubId);
-
-  if (categories.isLoading || clubDetail.isLoading) {
-    return <section>Loading...</section>;
+  if (isLoading || isError) {
+    return <></>;
   }
 
-  if (categories.isError || clubDetail.isError) {
-    return <section>Error...</section>;
-  }
-
-  return (
-    <Templates.ClubDetail>
-      <section>
-        <h1>club Detail</h1>
-      </section>
-    </Templates.ClubDetail>
-  );
+  return <Pages.ClubDetail />;
 };
 
 export default Club;
